% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clalc_Release.R
\name{calc.release}
\alias{calc.release}
\title{Calculate Reservoir release}
\usage{
calc.release(release.th, storage.th, storage, inflow = NA,
  inflow.th = NA, stor.max, dt = 1)
}
\arguments{
\item{release.th}{vector of release thresholds units, thesee can be calculated with \code{\link{monthlypctl.fromday}} [l3/t]}

\item{storage.th}{vector of storage thresholds, these can be calcualted with \code{\link{monthlypctl.fromday}}  [l3]}

\item{storage}{current storage [l3]}

\item{inflow}{current inflow, defalut is NA. (see inflow.th, this parameter is only used if inflow.th!=NA) [l3/t]}

\item{inflow.th}{the storage threshold above which inflow should be considered when calcualting outlfows
NA value indicates that inflows are never included (default is NA)}

\item{stor.max}{maximum storage capacity [l3]}

\item{dt}{timestep [t]}

\item{stor.min}{dead pool storage cpacity [l3]}
}
\value{
The simulated outlfow rate from the reservoir [l3/t]
}
\description{
Function to calculate reservoir release given current storage and operating thresholds
Assumes that the dead pool storage and release values are the first in the release pctl and storage pctl vectors
}
\examples{
calc.release(release.th=c(0,50,75,90), storage.th=c(100,200, 600, 1000),  storage= 500, inflow=20, stor.max=1200, dt=0.1, )

}
\seealso{
\code{\link{monthlypctl.fromday}}
}
