% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_MonthlyPercentiles.R
\name{get.monthlypctl}
\alias{get.monthlypctl}
\title{Get monthly Percentiles}
\usage{
get.monthlypctl(monthlydata, percentile.list, year.range)
}
\arguments{
\item{percentile.list}{A vector of the percentile values to be calcualted.}

\item{year.range}{The range of years from the original input data that should be used for calculation (firstyear, last year).
NOTE: If the beginning of the range is before the first year of the data it will default to starting with the first year of data and similarly if the
last year is after the end of the data it will default to the last year available.}

\item{monthly.data}{a matrix of values where the first column is the year and the next 12 columns are monthly values for that year.}
}
\value{
a matrix of the percentile values with 12 columns (one for each month) and a row for each percentile value given in percentile.list
}
\description{
Calculates the monthly percentiles for a matrix of monthly values give a range of years
}
\examples{
get.monthlypctl(monthlydata=monthmatrix, percentile.list=c(0.25, 0.5, 0.75), year.rang3=c(1950, 2000))

}
