% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Monthly.R
\name{make.monthly}
\alias{make.monthly}
\title{Make Monthly Matrix}
\usage{
make.monthly(dates, values, fun, na.val = NA)
}
\arguments{
\item{dates}{A vector of dates, these must be formatted as dates}

\item{values}{A vector of values associated with the dates, this vector must be the same length as the date vector}

\item{fun}{The name of the function to use for aggreagating values. Choices are mean, sum, first and last. Mean and sum return the average and total values for the month
while first and last return the values on the first an last days of the month respectively. Default is mean.}

\item{na.val}{na value for missing data in the timeseries. Defaults is NA.}
}
\value{
\enumerate{
\item mat: A matrix with 13 columns, column 1 is the year and columns 2-13 are the monthly aggregated values
\item ts: A xts time series of monthly values column1 is the date, column 2 is the monthly aggregated value
}
}
\description{
Function to aggregate daily timeseries to monthly values. Aggregation opptions are (1) average, (2) sum,
(3) first value of the month, (4) last value of the month. Required inputs are vectors of dates and daily values that
will be converted. The function returns a times series of monthly values and a matrix with one column per month.
}
\examples{
make.monthly(dates=datevec, values=atributevec, fun="mean", na.val=-999)

}
