% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_TimeSeries.R
\name{make.xts}
\alias{make.xts}
\title{Make an xts time series}
\usage{
make.xts(dates, values, fun, na.val = NA, na.omit = T, na.fill = F)
}
\arguments{
\item{dates}{A vector of dates, these must be formatted as dates}

\item{values}{A vector of values associated with the dates, this vector must be the same length as the date vector}

\item{na.val}{na value for missing data in the timeseries. Defaults is NA.}

\item{na.omit}{Option for whether NA values shoudl be omitted from the final time series or not. Default is True.}
}
\value{
An xts time series of monthly values column1 is the date, column 2 is the monthly aggregated value
}
\description{
Function to make an xts time series from vectors of dates and values removing all dates with NA observations.
}
\examples{
make.xts(dates=datevec, values=atributevec,  na.val=-999, na.omit=T)

}
