% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_MonthlyPercentiles_FromDaily.R
\name{monthlypctl.fromday}
\alias{monthlypctl.fromday}
\title{Get monthly Percentiles from daily}
\usage{
monthlypctl.fromday(timeseries, percentile.list, start = "1940-01-01",
  stop = "1950-03-01")
}
\arguments{
\item{timeseries}{a time series of daily values to calcualte percentiles from. This must be formatted as an xts time series. See \code{\link{make.xts}}.}

\item{percentile.list}{A vector of the percentile values to be calcualted.}
}
\value{
a matrix of the percentile values with 12 columns (one for each month) and a row for each percentile value given in percentile.list
}
\description{
Calculates the monthly percentiles of flow from a time series of daily values
}
\examples{
get.monthlypctl(monthlydata=monthmatrix, percentile.list=c(0.25, 0.5, 0.75), year.rang3=c(1950, 2000))

}
\seealso{
\code{\link{make.xts}}
}
