% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_releases.R
\name{res.sim.day}
\alias{res.sim.day}
\title{Simulate daily reservoir releases}
\usage{
res.sim.day(release.pctl, stor.pctl, inflow.ts, stor.init, stor.max,
  stor.min, sim.start, sim.stop, nsteps = 1, dt = 1)
}
\arguments{
\item{release.pctl}{a matrix of release thresholds
with 12 columns (one for each month) and a row for each percentile threhold
(see \code{\link{monthlypctl.fromday}} [l3/t])}

\item{stor.pctl}{a matrix of storage thresholds a matrix of the percentile values
with 12 columns (one for each month) nd a row for each percentile threhold
(see \code{\link{monthlypctl.fromday}} [l3])}

\item{inflow.ts}{an xts timeseries of inflows (see \code{\link{make.xts}},
note that NA's should be filled so that there are no missing days in the inflows) [l3/t]}

\item{stor.init}{an initial storage value for the reservoir [l3]}

\item{stor.max}{the maximum storage capacity of the reservor [l3]}

\item{stor.min}{the minimum usable storage capacity of the reservoir (i.e. the dead pool storage) [l3]}

\item{sim.start}{date to start simulation on (should be a string "yyyy-mm-dd")
(NOTE - if this is before the start date of the inflow time series simulation will by adjusted to start at the start of the inflows)}

\item{sim.stop}{date to end simulation on (should be a string "yyyy-mm-dd")
(NOTE - if this is after the end date of the inflow time series simulation will by adjusted to end at the last of the inflows)}

\item{nsteps}{the number of simulation steps to take within one timestep of the inflow timeseries, default is 1
(i.e. how often reservoir releases will be adjusted within a timestep)}

\item{dt}{the length of a time step in the inflow timeseries, default is 1.
(e.g. if you are providing daily inflows and the unts are m3/hour the dt would be 24, if the unts are m3/s it would be 86400 )}
}
\value{
returns a matrix of reservoir operations with four columns (1) date, (2)reservoir inflow [l3/t], (3) reervoir release [l3/t],
 (4) storage at the end of the day [l3]. The lenth (l) and time (t) units of the ouput will be the same as the inflow timeseries provided.
(NOTE even if your nstep is >1 this will return the aggreageted operations at the same time step as the inflow timeseries)
}
\description{
Simulate releases given and operating  policy
}
\examples{
calc.release(release.pctl, stor.pctl, inflow.ts, stor.init=100, stor.max=10000, stor.min=2, sim.start="1940-01-01",
sim.stop="1950-03-01", nsteps=240, dt=24) )

}
\seealso{
\code{\link{monthlypctl.fromday}} \code{\link{calc.release}}
}
